+++ CHANGELOG.rst
@@ -2,7 +2,13 @@ Only major changes are reported here. Releases that only bump the patch part of
 the version number (i.e. the number after the 2nd dot) are generally not listed
 here unless they include a fix to a specific issue reported on github.
 
-* 6.1.0:
+* 7.0.0:
+
+  - Breaking change: ``Event`` object no longer participate in the ad-hoc
+    hashability protocol that this library uses.
+
+  - ``EventSet`` are now implemented using ``Set`` for backing store instead of
+    an array.
 
   - New feature: introduce a pair of "compact" events named
     ``attributeNameAndValue`` and ``startTagAndAttributes`` which allow passing